refactor: divide product categories into compoments
